Fix the Emacs regex used for highlighting TS regexes.
The previous regex would treat // as a regex, which is wrong. In JS
(which TS follows to the letter, in this case), // is never a regular
expression. It *looks* like an empty regular expression, but if it
appears outside a string or comment, then it is the start of a single
line comment, not an empty regular expression. The old regex also had
trouble dealing with an escaped forward slash appearing in a
regex. The commit fixes both problems, and it adds tests to check for
such issues.