Use syntax-propertize-function to fontify regexps.
The complexity of regular expressions taxes the basic facilities for
funtification. We reuse the latest js.el code that handles
fontification to fix lingering issues with how typescript-mode has
handled regexp fontification so far.

We do not simply set `syntax-propertize-function` to
`js-syntax-propertize` because the latest version of js.el shipped
with Emacs has a buggy implementation. Our choices are: copy from the
latest version of js.el, come up with some other implementation, or
accept to live with buggy fontification.

Closes #44.