Drop typescript-parent-mode.
Fixes #75.

Copying some of the discussion from the issue thread:

"If I run `git blame` on the code, I see that setup dates from the
initial commit and so was probably inherited from `js-mode`. If I look
at the current `js-mode` bundled with my Emacs, however, there's no
such equivalent indirection there. Searching through Emacs's commits,
I see `prog-mode` was added to Emacs 24.1 (which dates from 2012
according to Wikipedia). So I guess if `typescript-mode` is to be used
with versions of Emacs *older than that*, it needs to be able to fall
back on `fundamental-mode`, but do we need to support versions that
old?"

(No, we don't.)