Run emacs --version prior to running cask.
If cask fails due to a installation issue with Emacs, cask does not provide good
diagnosis. If we try running Emacs prior to cask, at least whatever diagnosis is
provided by the OS (e.g. "cannot find such and such library") is going to be
readily readable in the test log.