Fix infinite loop in indentation.
This fixes an issue that was introduced in #126

In a case like this:

```
if (...) {
  .bleh();
}
```

running indentation on the line with `.bleh()` would (prior to this commit)
result in an infinite loop. The solution is to treat the dot as not being an
expression continuation. It is the start of the expression, not the continuation
of an expression already started earlier.