Add tree-sitter support for both tree-sitter implementations
There are for the time being two tree sitter implementations.  The first one is
the older, rust based variant written by @ubolonton which is widely used in the
emacs community, and the other is the newer variant from emacs proper, which for
the time being is situated in the feature/tree-sitter branch.